requirements.txt
openai ~= 1.14.3
python-dotenv ~= 1.0.1
bs4 ~= 0.0.2
openai == 1.14.3
google-generativeai == 0.4.1
streamlit == 1.32.2